Consolidate card details display and improve deck check UI
- Extract common CardDetails component for shared card display logic
- Combine best features: deck check "well" text, card search links/colors
- Replace text status indicators with Ban and MessageCircleQuestion icons
- Add full-width table borders between card rows for better separation
- Rename ExpandableCardRowSimple to ExpandableCardRow
- Improve visual hierarchy and consistency across Card Search and Deck Check
